requirements.txt
streamlit
pandas
numpy
statsmodels
websockets
python-dateutil
